// src/lib/schemas/orderPublicResponseDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * 5Sport API
 * API documentation for 5Sport authentication and services
 * OpenAPI spec version: 1.0
 */
import type { OrderPublicResponseDtoPaymentStatus } from './orderPublicResponseDtoPaymentStatus';

export interface OrderPublicResponseDto {
  /** Mã đơn hàng */
  orderCode: string;
  /** Họ */
  lastName: string;
  /** Tên */
  firstName: string;
  /** Trạng thái thanh toán */
  paymentStatus: OrderPublicResponseDtoPaymentStatus;
  /** Mã giao dịch thanh toán */
  paymentId?: string;
  /** Thời gian thanh toán thành công */
  paidAt?: string;
}