// src/lib/schemas/orderPublicResponseDtoPaymentStatus.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * 5Sport API
 * API documentation for 5Sport authentication and services
 * OpenAPI spec version: 1.0
 */

/**
 * Trạng thái thanh toán
 */
export type OrderPublicResponseDtoPaymentStatus = typeof OrderPublicResponseDtoPaymentStatus[keyof typeof OrderPublicResponseDtoPaymentStatus];


export const OrderPublicResponseDtoPaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  CANCELED: 'CANCELED',
  REFUNDED: 'REFUNDED',
} as const;